# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "园智汇 · AI 产业园区智慧运营管理平台",
  "description": "SmartPark AI Operations Platform\n\n面向产业园区集团的多租户智慧运营平台，包含数据驾驶舱、空间资产、企业服务、招商运营、合同财务、项目管理、物业运维、安全管理、审批中心、AI Agent（1 个主 Agent + 11 个子 Agent）、数据中心、报表中心、日志审计与系统管理。\n\n**数据说明**：演示环境，全部数据为演示数据（is_demo=True），由数据库实时聚合计算。",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/approvals": {
   "get": {
    "tags": [
     "审批中心"
    ],
    "summary": "List Approvals",
    "operationId": "list_approvals_api_approvals_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "approval_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Approval Type"
      }
     },
     {
      "name": "only_mine",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Only Mine"
      }
     },
     {
      "name": "keyword",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Keyword"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Approvals Api Approvals Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/approvals/todo": {
   "get": {
    "tags": [
     "审批中心"
    ],
    "summary": "My Todo",
    "description": "我的待办审批（按当前审批人/角色匹配）。",
    "operationId": "my_todo_api_approvals_todo_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response My Todo Api Approvals Todo Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/approvals/{approval_id}": {
   "get": {
    "tags": [
     "审批中心"
    ],
    "summary": "Approval Detail",
    "operationId": "approval_detail_api_approvals__approval_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "approval_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Approval Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Approval Detail Api Approvals  Approval Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/approvals/{approval_id}/decision": {
   "post": {
    "tags": [
     "审批中心"
    ],
    "summary": "Decide",
    "description": "审批决策（逐级推进；全部通过才最终批准）。",
    "operationId": "decide_api_approvals__approval_id__decision_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "approval_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Approval Id"
      }
     },
     {
      "name": "decision",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Decision"
      }
     },
     {
      "name": "comment",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Comment"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Decide Api Approvals  Approval Id  Decision Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/approvals/{approval_id}/withdraw": {
   "post": {
    "tags": [
     "审批中心"
    ],
    "summary": "Withdraw",
    "description": "撤回审批（仅申请人本人或集团管理员）。",
    "operationId": "withdraw_api_approvals__approval_id__withdraw_post",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "approval_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Approval Id"
      }
     },
     {
      "name": "reason",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Reason"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Withdraw Api Approvals  Approval Id  Withdraw Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/login"
     }
    }
   }
  }
 }
}